Add name validation for projects
Currently, even projects with empty names are allowed. The UI
tries to avoid this by using a required attribute on the inputs,
but it's trivial for users to work around that.

To not break existing projects with invalid names, only validate
for new projects, and when the project's name has changed.

For the validation format, use a case-insensitive version of
Validation::ALLOWED_NAME_PATTERN.