Allow resource creation if the project has credit
We expect our customers to add a payment method to create resources,
which helps us prevent fraud.

However, sometimes customers want to pay by bank transfer and prefer not
to add a payment method.

We should allow the creation of new resources if the project has credit.
But we need to create billing information for the customer without
payment method since we require it for invoicing.