Exclude uncommitted cache entries when cleaning up cache storage
In the caching workflow, the client first reserves the cache and then
commits it.

The size of the cache entry is set when committing the cache in certain
cases.

If we clean up the cache storage while a cache entry is reserved but not
committed, it fails with the following error because the size is nil:

    TypeError: nil can't be coerced into Integer