Return cache in incoming key order for partial match
We sorted cache entries according to the order of the incoming keys and
returned the first match. However, this only works when the full key is
provided. It seems that partial match doesn't follow the same ordering.
Instead, it simply orders the entries by their creation time. We should
order them by the incoming key order.

I used similar logic to the full match for this. The only difference is
that I used `find_index` and `start_with?` instead of `index`.