Adding burstable family
Adding a burstable VM family with an ability to place multiple VMs in once slice. The allocator is updated to select VMs that may be fully used but have some space in existing slices and the VmHostSliceAllocator decides if an existing slice can be reused or a new one needs o be created, for each host. The hosts where an existing slice can be used are preferred.

Tests updated and passing 100% with full coverage.

This is no finished yet :)