Improved health checks for inference endpoints
The latest version of AI base images support a new /health endpoint. In
contrast to the /up endpoint, which succeeds if the inference gateway is
operational, the /health endpoint performs more thorough checks and
fails if either inference gateway or inference engine are not available.

This commit switches inference endpoint load balancers from the /up
to the /health endpoint.