Create pages for unavailable inference endpoints
A new label `unavailable` is added the the inference endpoint replica nexus.
When a replica becomes unavailable, it is moved to that label. In that state,
a page is created and the prog checks the replica's availability status more
frequently. When the replica becomes available again, it hops out of
`unavailable` and resolves the page.

An operator can set the `maintenance` semaphore on an inference endpoint,
which suppresses paging on replica unavailability.