Append docker IPv6 config to daemon.json instead of overwriting it
We found that Docker doesn't work smoothly with IPv6 by default. To
address this, we started to configure Docker to work correctly with IPv6
at d98fd5c9.

But it overwrites the existing configuration. This is not good because
it may break the existing configuration. For example our GPU runners
have an existing configuration for the nvidia runtime.