Have invalid project name error message explain why
Currently, the error message is the default "is invalid", which
isn't that helpful to the user.

This is a bit explicit and technical, in that it mentions ASCII,
but without ASCII, it implies that non-ASCII letters would be
acceptable, which is incorrect.  I would guess that our users
are technical enough to understand this message.

Co-authored-by: Benjamin Satzger <53236451+bsatzger@users.noreply.github.com>