Use the static_cache plugin in the ActionType table
The ActionType is a static table that is not updated at runtime.
For these types of tables, using the static_cache plugin is
preferred, as it can reduce queries by using the static cache.

Use the static_cache_cache plugin globally so that the entries for
static_cache tables are cached just like the schema and indexes
are cached, avoiding a query per static_cache table to get the
values.