Adding overcommiting and a Basic family
Adding a new family of VMs - Basic. Basic VMs share a VmHostSlice and multiple of them run in once slice. They allow for overcommit, meaning that more VMs can be placed inside the slice than there are actual CPU resources. For now the overcommit factor is 2x. This allows to insert 8 1-cpu VMs (Basic-1) or 4 2-cpu VMs (Basic-2) into a a 2-cores (or 4-cpu) slice.
This commits also decouples the cores and cpus values in Vm and VmSize. The cores designate how many VmHost cores are reserved for a slice hosting the Vm, and therefore removed from the host. The cpus designate how many vcpus (threads) are given to the Vm through the hypervisor. For "standard" family those two were always related through the architecture, but for "basic" family they are not directly mapped. This required changes to the cloud_hypervisro_cpu_topology.

Basic VMs, as specified in VmSize structure, allocate a 2-core slice and place 1-cpu or 2-cpu Vms into it.