Relax private subnet deletion check
When processing a request to delete a PrivateSubnet, the system
currently verifies whether the subnet contains any VMs. If VMs are
present, a dependency error is raised. This behavior can lead to
unexpected errors if a client deletes all VMs and subsequently
attempts to delete the subnet.

To address this, the deletion check is being adjusted to allow the
subnet to be deleted even if it still contains VMs, provided all the
VMs are already marked for deletion. This ensures the subnet
deletion request succeeds in such scenarios.