Switch invited user policies to inviting to a specific subject tag
This has mostly the same effect as the previous behavior, though
a user would have to create a Member subject tag and grant it
appropriate permissions first.  We might do that automatically
add a Member subject tag automatically for new projects in the
future, but for now, only the Admin tag will be available.

This temporarily breaks changing policies for users, since the
policy_select component is used for both invited users and
current users.  Changing policies for users will be removed
soon, since it will replaced by a new access control section.