Remove hard dependency to MinIO connection in Postgres serializer
We need to connect to MinIO cluster to calculate earliest restore time, however
this is not always possible and can error out due to various reasons. In such
cases, instead of returning 500, it is better to just hide restore UI. This
commit does that by surrounding the code that tries to connect MinIO with a
rescue block.