Ensure request body is rewindable
Committee requires rewindable input.  Input is not guaranteed to
be rewindable in Rack 3, and Rack::Lint will complain about if
you call rewind on input.

To work around this, use Rack::RewindableInput in development mode,
which is designed to handle rewinding.

To make sure Clover is compatible with the current Rack version in
use, use Rack::Lint when running in test mode, unless SimpleCov is
defined.  Since the default rake task and CI test both with and
without coverage, this ensures tests will cover both the Rack::Lint
and non-Rack::Lint cases.