Fix Authorization.all_permissions to handle NULL action_id
If there is a matched policy with a NULL action_id, that means all
actions are allowed, so all action type names should be returned.

Switch Clover#has_project_permission back to using
@project_permissions now that it works.  There are often multiple
calls to has_project_permission for one request, and we should
avoid a separate query for each.