Support nested subject/action/object tag inclusion via recursive CTEs
This may be easiest to explain via an example, and we'll use object
tags as an example (subject and action tags work similarly).

For background, assume a project has the following set of nested tags:

* Servers
  * Web Servers
    * NA Web Servers
    * EU Web Servers
    * AP Web Servers
  * NTP Servers
  * DNS Servers

If a server in the "EU Web Servers" tag, it should also be considered
in the "Web Servers" and "Servers" tag.  If a user is granted access
to the "Web Servers" tag, it should be able to access objects with the
"EU Web Servers" tag.

Before this commit, if an object was in the "EU Web Servers" tag, it
could not be managed by a user with access to the "Web Servers", since
tag lookup was only a single level deep.  This changes the query to
use recursive common table expressions, to allow for arbitrary depth
(depth is currently limited to 32 levels by default, though this can
be changed via an environment variable).

For the main access query (matched_policies_dataset), the CTE goes
from deepest level to main level, as we find to find the tags directly
including the object, and then move back to the root tags:

1. Look for object X, found in "EU Web Servers" (T1)
2. Look for applied object tags referencing "EU Web Servers", found in "Web Servers" (T2)
3. Look for applied object tags referencing "Web Servers", found in "Servers" (T3)
4. Look for applied object tags referencing "Servers", found none
5. Look for access control entries where object_id IN (X, T1, T2, T3)

For Dataset#authorize, there is a matched policies dataset query to get
all object_ids, and then a recursive query that goes from the root
through the tags to get all objects.

For example, if using the following in Clover:

  dataset_authorized(projects.vms_dataset, "VM:view")

Assuming there was a single matching access control entry for the
"Servers" tag, then the recursive query would do something like this:

1. Look for applied object tags where the tag is "Servers",
   found "Web Servers", "NTP Servers", "DNS Servers" (and any
   direct objects)
2. Look for applied object tags where the tag is "Web Servers",
   "NTP Servers", or "DNS Servers, found "NA Web Servers",
   "EU Web Servers", and "AP Web Servers" (and any direct objects)
3. Look for applied object tags where tag is "NA Web Servers",
   "EU Web Servers", and "AP Web Servers", find only direct
   objects in those.
4. Look for applied objec tags where the tag is one of the direct
   objects, finding none

Then it would do something equivalent to:

  projects.vms_dataset.where(id: any_found_tags_or_direct_objects)

Note that `any_found_tags_or_direct_objects` is likely going to
contain non-Vm IDs, both tags and other types of objects.  However,
as all IDs are UUIDs, this is fine, it will just ignore IDs that
do not match.

For all_permissions, there is a matched_policies_dataset subquery
to get all action_ids, and then a similar recursive query used for
Dataset#authorize that goes from the root through the tags to get
supported actions.