Make form/select component support optgroup
By providing a hash of options, the hash key specifies the optgroup,
and the values are options.

There was one case where a hash was already used, in the
form/country_select component, so I switched that to use a hash.

This allows each optgroup to also support attributes.
If there is an option with the same text in two different
optgroups in the same select tag (where the option values are
different), it's challenging to get Capybara to select the
desired option.  This allows you to provide different id
attributes to different optgroups, so that Capybara can select
options inside a specific optgroup.

This will be used by the upcoming access control editor.

This excludes the components/form/select view from erb_formatter,
because erb_formatter does not handle it correctly, complaining
about mismatched tags when there are none.