Make form/select component support optgroup
By providing a hash of options, the hash key specifies the optgroup,
and the values are options.

There was one case where a hash was already used, in the
form/country_select component, so I switched that to use a hash.

This will be used by the upcoming access control editor.