Add AccessControlEntry validation
This validates that subjects, actions, and objects in the access
control entry are validate and related to the current project.

In a few places in the specs where we are testing access control
entries, skip validation as it isn't important to what is being
tested.  In a couple other places, make small changes to work
with the validation.