Nap during database recovery if cannot connect to the database
The query to check if the database is in recovery mode can fail if the database
is not yet ready to accept connections due to ongoing recovery. In such cases,
instead of raising an exception, which in turn would cause retry backoff to
kick in, we should just sleep for a while and retry the query.