Fix referrer handling for repeated failures
The previous code only handled 2 failures in a row.  If you submitted
a form with more than 2 failures in a row, it would loose track of the
original referrer and break.

This removes the flag set by the internal request that the session
has already been serialized, so that the real request will have it's
session serialized with the correct flash set.

An upcoming commit will include a test that depends on this.