Add UI for tag management
This allows, for subject, action, and object tags:

* Creating tags
* Renaming tags
* Removing tags
* Adding members to tags
* Removing members from tags

The Admin subject tag is special, and cannot be renamed, removed, or included
in other tags.  It also cannot have all accounts removed from it, to ensure
there is at least one account with full access.

When adding members to a tag, the member:

* Cannot be the tag itself (trivial loop prevention)
* Cannot already be a member (database unique constraint would catch this)
* Cannot already include the receiver as a member, either directly or
  indirectly (loop prevention).
* Must be a valid member of the tag, by calling *Tag.valid_member?
* Cannot be the Admin subject tag, if receiver is subject tag.

I think the add member code for preventing loops would be subject to
race conditions, so this uses a serializable transaction for it.  I'm
not positive even that will prevent all loops, but it should make
creating loops very difficult.  We do have a configurable recursion
limit in all recursive queries, so loops are not catastrophic.

This adds the same validation we have on project names to tag names.
It also removes any references to the tag when deleting it:

* Removes memberships in tag itself
* Remove memberships of current tag in other tags
* Removes any access control entries directly on the tag itself

Just as with access control entries, when showing tag membership,
tokens are hidden. This may only affect the Admin subject tag, as
that is the tag that tokens will be added to by default.

This uses the same routing tree and views to manage all tag types,
which significantly DRYs things up.