Add UI for tag management
This allows, for subject, action, and object tags:

* Creating tags
* Renaming tags
* Removing tags
* Adding members to tags
* Removing members from tags

The Admin subject tag is special, and cannot be renamed, removed, or included
in other tags.  It also cannot have all accounts removed from it, to ensure
there is at least one account with full access.

Just as with access control entries, when showing tag membership,
tokens are hidden. This may only affect the Admin subject tag, as
that is the tag that tokens will be added to by default.

This uses the same routing tree and views to manage all tag types,
which significantly DRYs things up.