Add commented out RSpec configuration for debugging truncated output issues
When RSpec truncates expected and actual output for user display,
it does not check that the truncated output is different.  This
leads to difficult to debug issues, since you have no indication
where the difference is.  This allows for easier debugging of
the issues, by temporarily not using truncated output.