Make authorization code check that object_id is related to project_id
This check is in general redundant with other checks made in Clover,
but I think for defense in depth, it is worth having.  In case the
object retrieval is not correctly filtered to the given project,
this will prevent access as long as it attempts to authorize.

This changes the SecureRandom.uuid calls in the authorization spec
to force them to accidentally match a recognized UBID class, as
otherwise the authorization code would call filter_authorize_dataset
on a class that doesn't implement it.