Add a migration for an index for access_tag.{hyper_tag_id,project_id}
This index should allow use of an index-only scan for the subquery added
by Authorization::HyperTagMethods.filter_authorize_dataset.

This is added in a separate migration, because it uses CREATE
INDEX CONCURRENTLY, and that is not supported inside transactions
(and migrations use transactions by default).  I'm not sure, but
I'm guessing the access_tag table size is large enough we would
want to create the index concurrently to avoid blocking access
to the table.