Replace Postgres:Firewall:{view,edit} permissions with Postgres:{view,edit}
Postgres firewalls are not a separate object type.  There is
PostgresFirewallRule, but the Postgres:Firewall permissions
are tested against the PostgresResource, so for consistency,
it makes sense to keep the type:action format used by other
permissions, and not support colon as part of the type. After
further discussion, there is agreement that separate permissions
for Postgres firewall rules are not needed, so using the normal
Postgres view/edit permissions are fine.

While here, avoid N+1 fwedit access check queries in the
view/postgres/show template, where N is the number of firewall
rules.