Add ObjectMetatag
The ObjectMetatag class is being introduced to differentiate between
having access to objects in a tag, versus having access to the tag
itself.  Consider an access control entry such as:

  project_id: project.id
  subject_id: user.id
  action_id: nil # all actions
  object_id: object_tag.id

The desired access here is the user has full access to all members
of the tag.  However, if you do not differentiate access to members
of the tag from access to the tag itself, this is equivalent to
admin access, because it would give the user the ability to add
any objects they want to the tag itself, which would be equivalent
to allowing full access to the entire project.

By separating an object tag from its metatag, you can avoid this
issue.  Granting access to an ObjectTag will grant you access to
members of that Tag.  Granting access to an ObjectMetatag will
grant you access to the tag itself, but not grant access to members
of the tag.

To avoid having to store a separate uuid for ObjectMetatag, we
shift the UBID, so it starts with t2 instead of t0, but is
otherwise the same (t2 instead of t1 so that parity remains
the same).

ObjectMetatag is a DelegateClass for ObjectTag, so that you can
more or less treat an ObjectMetatag as an ObjectTag, except that
the id method returns the metatag uuid and the ubid method returns
the metatag ubid.

This will be used in an upcoming commit that supports granular
access control to the access control features (i.e. allowing user
A to add/remove members only to object tag B).