Make ObjectTag Dataset#authorize handle metatag lookup
When using ObjectTag.dataset.authorized, access should only be
allowed if the metatag for the object tag is allowed, because
that is what allows access to the tag itself.  Access should
not be allowed just because the tag is allowed, as otherwise,
any user with access to an object tag containing say specific
VMs, to change the members of the object tag to elevate their
access.

Doing this requires converting from metatag uuid to tag uuid
using PostgreSQL string functions.  This converts metatag uuids
to tag uuids so the metatag lookup works, and converts tag uuids
to uuids that would be invalid ubids, ensuring they shouldn't
match the uuid for an existing tag.