Remove the vcpu count check in Vm::params_json
We assume in the code that in all x64 hosts `:total_cpus ==
2*:total_cores`, but this isnot the case for GEX hosts, which we model
as `:total_cores=>14, :total_cpus=>14`.

Because of this, the number of `max_vcpus` that we calculate in
`Vm::params_json` is equal to half the number of actual vcpus requested,
so the recent consistency check that we added raised an error.

Since the inconsistency is a bug from past, in this PR we remove the
check to not raise exceptions in prod (i.e. behaviour will be as it used
to be), and then we will fix the actual issue in a future PR.