Add token access control section
By default, tokens are added to the Admin subject tag and have full
access. This allows you to click on the token ID in the token list,
and opt to restrict the token.  When the token is restricted, it
defaults to no access, and you have to grant the access you want
the token to have, using a similar interface as the current access
control entry section, except without a subject (since the subject
is always the token itself).

I was originally thinking of merging this into the existing access
control route, but it would have required adding a lot of conditionals
and complexity, and I'm not sure if would save that much code in the
long run.  This does not reuse the existing template for viewing
access control entries, but it does reuse the existing template for
creating/editing access control entries.

This adds ApiKey#unrestricted_token_for_project? and
ApiKey#restrict_token_for_project methods for handling the restriction.
These currently don't check that they are called on a personal access
token and not an AI interface token, but misuse seems unlikely.

Add a flash notice when deleting a token.

Fix breadcrumb when creating/updating access control entry.