DRY up code for {Subject,Action,Object}Tag members
There was previously duplicate code in the access control
entry form and the tag membership form.  This centralizes
the options as class methods on the tag classes, and
has both forms call the methods to use the same
options.

The access control entry form was missing the the
{Subject,Action,ObjectTag} options for the object_id value.
Now that behavior is centralized, it picks those options
up.

This caused breakage in the Capybara specs, because
previously unambiguous option texts became ambiguous.
Add optgroup ids to differentiate Object tags from
metatags in the object_id select box.  Use the Capybara
within method in the specs with the correct ids to ensure
unambiguous access.

Simplify the tag membership page, avoiding the case
expression and the lambda.