Have {Subject,Action,Object}Tag#destroy remove metatag entries
SubjectTag and ActionTag can be stored as the object_id in both
applied_object_tag and access_control_entry. The metatag for
ObjecTag can also be stored in the same place.  This ensures
all related entries are removed.

Add ObjectMetagtag.to_meta_uuid to handle the need in the specs
to convert ObjectTag id to metatag id.