Show Project as an object tag member option
ObjectTag#valid_member? already considered the project as a valid
member, but the UI didn't expose this.

Originally, I wasn't sure this was needed, as each access control
entry already has a project id associated with it.  However, if
you consider an access control entry such as:

  subject_id: account A
  action_id: nil (all actions)
  object_id: tag B

Then the expectation is that the user should not be able to take
project actions on the project unless the project is in tag B.
In this case, it's necessary to either put the project in the tag
if the account should have access, or not put it in the tag if it
shouldn't have access.