Add {Subject,Object}Tag::Cleanup modules
These can be included in other classes so the related tag
memberships and access control entries are deleted when the
object is destroyed.

It's possible to do this by defining associations and using
association_dependencies, but defining unnecessary associations
is a bad idea from a memory usage standpoint, and this approach
should be simpler and more efficient.