Add a converter from access policy entries to access control entries
A project can have multiple access policies, and each access policy
can have multiple entries.  Each entry in an access policy can
convert into an access control entry.  If there are multiple
subjects, actions, or objects in the access policy entry, a tag
can be created for each, and that tag can be used in the access
control entry.

This breaks the process down into three separate methods:

* parse_access_policy_rows: parses the entries into arrays of uuids,
  one for each access control entry.  The subject, action, and
  object entries in each array can themselves be arrays of uuids.

* convert_access_policy_rows: calls parse_access_policy_rows,
  and further converts the entries, returning an unsaved
  AccessControlEntry for each.  The subject_id, action_id, and
  object_id values, if they need to be tags, are arrays, with the
  unsaved tag as the first element, and an array of checked member
  uuids as the second element.

* save_converted_access_policy_rows: calls convert_access_policy_rows,
  and saves each access control entry.  If the access control entry
  references tags, saves the tags first.

For all three methods, they return a hash with :aces and :failures
keys.  The :aces key values are described above.  The :failures
key is an array of arrays of failure information, which can be used
for debugging.

No specs for this yet.